use an event, insertion logic own class
